Enemy spawner code, Enemy info script
Enemy spawning code added, EnemyInfo can be defined in the enemy prefabs, pathfollow now gets the speed and spawnpoint from the EnemyInfo

FindSpawnpoint.cs has a class with functions to return an array of spawnpoint GameObjects or an array of the positions of the spawnpoints. This is used in Spawner.cs and PathFollow.cs and can be used in other scripts that require the spawnpoints.