Purify Turrets and Buff Drops
Changed EnemyInfo in order to account for the purity hp. Each time an enemy dies from a purify turret, it will drop a random item. Added 3 new prefabs for said items that. That we wish to add. 2 new scripts have been saved in Enemy Scripts, "Loot" and "LootBag". Loot simply has all of the loots that we have created with prefab and LootBag will choose which loot from Loot to drop.  To create a new piece of Loot, right click in the project menu and go to create->Loot. You then add in the sprite, chance that it will drop, and weight (For dramatic effect of throwing it somewhere)